requirements.txt
pytest==8.3.3
selenium==4.25.0
anthropic==0.34.2
python-dotenv==1.0.1